+++ StockApp/app/src/main/java/com/example/stockapp/totalcard/TotalData.kt
@@ -3,6 +3,6 @@ package com.example.stockapp.totalcard
 import java.math.BigDecimal
 
 data class TotalData(
-    var amount : BigDecimal,
-    var amtChange: BigDecimal
+    var amount : Double,
+    var amtChange: Double
 )
